Refactor profile library logic out of app.tsx into custom hook
